address review: move validator to codechecker_common.util
Per #4830 review feedback: is_valid_postgresql_db_name belongs with
the other generic helpers in codechecker_common.util, not in the web
server's routing module. The function and its tests are unchanged;
only the import paths in product_server.py and test_request_routing.py
are adjusted.
